requirements.txt
torch
tokenizers
espnet
soundfile
sounddevice
scipy
requests
playsound
PyGObject
